+++ src/parser.js
@@ -43,9 +43,9 @@ var grammar = {
             ['::self', 'return "SELF";'],
             ['[0-9]+(?:\\.[0-9]+)?\\b', 'return "NUMBER";'], // 212.321
             ['"(?:\\\\.|[^"])*"', 'return "STRING";'],       // "foo" "with \" escaped"
+            ["'(?:\\\\.|[^'])*'", 'return "STRING";'],       // 'foo' 'with \' escaped'
             ['/(?:\\\\.|[^/])+/i?', 'return "REGEXP"'],      // /foo/i
             ['[a-zA-Z_][a-zA-Z_$0-9]*', 'yytext = JSON.stringify(yytext); return "SYMBOL";'], // foo123
-            ["'(?:\\\\.|[^'])*'", 'return "SYMBOL";'],       // 'some-symbol' 'with \' escaped'
 
             // comment
             ['{ows}//.*?(\\n|$)', '/* a comment */'],