+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added computed property names support in object literals
 - Added array-like notation to access properties, e.g. `foo["bar"]` or `array[1]`
+- Changed single quoted string meaning to treat it as a string (used before for a property access with prohibited characters in name)
 - Fixed scope issue for method arguments, a scope the same as for query root (#1)
 
 ## 1.0.0-alpha.4 (November 25, 2018)